feat(UserProfilePage): add adaptive widgets on UserProfilePage and clean code
- Add AppBarWithBanner widget
- Add UserBasicInfo widget
- Add CTAButton widget
- Add Friend page layout for UserProfilePage
