+++ lib/contacts/user_profile_page/models/models_export.dart
@@ -1,0 +1,5 @@
+export 'app_bar_with_banner.dart';
+export 'cta_button.dart';
+export 'network_profiles_body.part.dart';
+export 'profile_data_body.part.dart';
+export 'user_basic_info.dart';
