+++ lib/contacts/user_profile_page/models/profile_data_body.part.dart
@@ -1,5 +1,3 @@
-import 'dart:ffi';
-
 import 'package:flutter/material.dart';
 import 'package:ui_library/ui_library_export.dart';
 import 'package:uplink/contacts/user_profile_page/user_profile_friends_page.dart';
